requirements.txt
streamlit>=1.28.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
reportlab>=4.0.4
mysql-connector-python>=8.1.0
python-dotenv>=1.0.0
